+++ NEWS.md
@@ -1,5 +1,7 @@
 # parsnip (development version)
 
+* Improved errors in cases where the outcome column is mis-specified. (#1003)
+
 # parsnip 1.1.1
 
 * Fixed bug where prediction on rank deficient `lm()` models produced `.pred_res` instead of `.pred`. (#985)